+++ MasterDataFlow.Intelligence/Genetic/GeneticFloatCommand.cs
@@ -6,14 +6,14 @@ using System.Text;
 namespace MasterDataFlow.Intelligence.Genetic
 {
     [Serializable]
-    public class GeneticFloatDataObject<TGeneticFloatItem> : GeneticDataObject<TGeneticFloatItem, int>
+    public class GeneticFloatDataObject<TGeneticFloatItem> : GeneticDataObject<TGeneticFloatItem, float>
         where TGeneticFloatItem : GeneticFloatItem
     {
 
     }
 
     [Serializable]
-    public abstract class GeneticFloatItem : GeneticItem<int>
+    public abstract class GeneticFloatItem : GeneticItem<float>
     {
         protected GeneticFloatItem(GeneticItemInitData initData)
             : base(initData)
@@ -22,7 +22,7 @@ namespace MasterDataFlow.Intelligence.Genetic
     }
 
     public abstract class GeneticFloatCommand<TGeneticFloatDataObject, TGeneticFloatItem>
-        : GeneticCommand<TGeneticFloatDataObject, TGeneticFloatItem, int>
+        : GeneticCommand<TGeneticFloatDataObject, TGeneticFloatItem, float>
         where TGeneticFloatDataObject : GeneticFloatDataObject<TGeneticFloatItem>
         where TGeneticFloatItem : GeneticFloatItem
     {
